+++ test/report/dot/folderLevel/index.spec.js
@@ -1,20 +1,22 @@
 const fs = require("fs");
+const path = require("path");
 const expect = require("chai").expect;
 const render = require("../../../../src/report/dot/folderLevel");
-const deps = require("./fixtures/dependency-cruiser-2019-01-14.json");
-const orphans = require("./fixtures/orphans.json");
-const rxjs = require("./fixtures/rxjs.json");
+const deps = require("./mocks/dependency-cruiser-2019-01-14.json");
+const orphans = require("./mocks/orphans.json");
+const rxjs = require("./mocks/rxjs.json");
 
+const mockFolder = path.join(__dirname, "mocks");
 const consolidatedDot = fs.readFileSync(
-  "test/report/dot/folderLevel/fixtures/dependency-cruiser-2019-01-14.dot",
+  path.join(mockFolder, "dependency-cruiser-2019-01-14.dot"),
   "utf8"
 );
 const consolidatedOrphansDot = fs.readFileSync(
-  "test/report/dot/folderLevel/fixtures/consolidated-orphans.dot",
+  path.join(mockFolder, "consolidated-orphans.dot"),
   "utf8"
 );
 const consolidatedRxJs = fs.readFileSync(
-  "test/report/dot/folderLevel/fixtures/consolidated-rxjs.dot",
+  path.join(mockFolder, "consolidated-rxjs.dot"),
   "utf8"
 );
