+++ test/report/dot/folderLevel/mocks/dependency-cruiser-2019-01-14.dot
