+++ test/report/dot/moduleLevel/mocks/defaultmodulecolor.dot
