+++ test/report/dot/moduleLevel/mocks/richmodulecolor.dot
