+++ test/report/dot/moduleLevel/mocks/richmodulecolor.json
