+++ test/report/dot/moduleLevel/mocks/unresolvable.dot
@@ -1,0 +1,22 @@
+strict digraph "dependency-cruiser output"{
+    ordering=out
+    rankdir=LR
+    splines=true
+    overlap=false
+    nodesep=0.16
+    ranksep=0.18
+    fontname="Helvetica-bold"
+    fontsize=9
+    style="rounded,bold,filled"
+    fillcolor="#ffffff"
+    compound=true
+    node [shape=box style="rounded, filled" fillcolor="#ffffcc" height=0.2 fontname=Helvetica fontsize=9]
+    edge [color="#00000033" penwidth=2.0 arrowhead=normal fontname=Helvetica fontsize=9]
+
+    "./not-at-home" [label="not-at-home" color="red" fontcolor="red" ]
+    subgraph "cluster_." {label="." subgraph "cluster_./this" {label="this" subgraph "cluster_./this/path" {label="path" subgraph "cluster_./this/path/does" {label="does" subgraph "cluster_./this/path/does/not" {label="not" "./this/path/does/not/exist" [label="exist" color="red" fontcolor="red" ] } } } } }
+    subgraph "cluster_test" {label="test" subgraph "cluster_test/cli" {label="cli" subgraph "cluster_test/cli/fixtures" {label="fixtures" subgraph "cluster_test/cli/fixtures/unresolvable-in-sub" {label="unresolvable-in-sub" "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" [label="refers-to-an-unresolvable-module.js" URL="STUFFCOMINGBEFORE/test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js"] } } } }
+
+    "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" -> "./not-at-home"
+    "test/cli/fixtures/unresolvable-in-sub/refers-to-an-unresolvable-module.js" -> "./this/path/does/not/exist"
+}
