+++ test/report/err/err.spec.js
@@ -2,12 +2,12 @@
 const expect = require("chai").expect;
 const chalk = require("chalk");
 const render = require("../../../src/report/err/err");
-const okdeps = require("../fixtures/everything-fine.json");
-const deps = require("../fixtures/cjs-no-dependency-valid.json");
-const warndeps = require("../fixtures/err-only-warnings.json");
-const erradds = require("../fixtures/err-with-additional-information.json");
-const orphanerrs = require("../fixtures/orphan-deps.json");
-const circularerrs = require("../fixtures/circular-deps.json");
+const okdeps = require("./mocks/everything-fine.json");
+const deps = require("./mocks/cjs-no-dependency-valid.json");
+const warndeps = require("./mocks/err-only-warnings.json");
+const erradds = require("./mocks/err-with-additional-information.json");
+const orphanerrs = require("./mocks/orphan-deps.json");
+const circularerrs = require("./mocks/circular-deps.json");
 
 describe("report/err", () => {
   let chalkLevel = chalk.level;
