+++ test/report/err/mocks/cjs-no-dependency-valid.json
