+++ test/report/err/mocks/everything-fine.json
@@ -1,0 +1,204 @@
+{
+    "modules": [
+        {
+            "source": "test/cli/fixtures/cjs/root_one.js",
+            "dependencies": [
+                {
+                    "resolved": "test/cli/fixtures/cjs/one_only_one.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./one_only_one",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "test/cli/fixtures/cjs/one_only_two.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./one_only_two",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "test/cli/fixtures/cjs/shared.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./shared",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "test/cli/fixtures/cjs/sub/dir.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./sub/dir",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "fs",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "fs",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "test/cli/fixtures/cjs/node_modules/somemodule/src/somemodule.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "somemodule",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "fs",
+            "followable": false,
+            "coreModule": true,
+            "couldNotResolve": false,
+            "dependencies": []
+        },
+        {
+            "source": "test/cli/fixtures/cjs/one_only_one.js",
+            "dependencies": [
+                {
+                    "resolved": "path",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "path",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "path",
+            "followable": false,
+            "coreModule": true,
+            "couldNotResolve": false,
+            "dependencies": []
+        },
+        {
+            "source": "test/cli/fixtures/cjs/one_only_two.js",
+            "dependencies": [
+                {
+                    "resolved": "path",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "path",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "test/cli/fixtures/cjs/shared.js",
+            "dependencies": [
+                {
+                    "resolved": "path",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "path",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "test/cli/fixtures/cjs/sub/dir.js",
+            "dependencies": [
+                {
+                    "resolved": "test/cli/fixtures/cjs/sub/depindir.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./depindir",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "path",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "path",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "test/cli/fixtures/cjs/sub/depindir.js",
+            "dependencies": [
+                {
+                    "resolved": "path",
+                    "coreModule": true,
+                    "followable": false,
+                    "couldNotResolve": false,
+                    "module": "path",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "test/cli/fixtures/cjs/node_modules/somemodule/src/somemodule.js",
+            "dependencies": [
+                {
+                    "resolved": "test/cli/fixtures/cjs/node_modules/somemodule/src/moar-javascript.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "./moar-javascript",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                },
+                {
+                    "resolved": "test/cli/fixtures/cjs/node_modules/somemodule/node_modules/someothermodule/main.js",
+                    "coreModule": false,
+                    "followable": true,
+                    "couldNotResolve": false,
+                    "module": "someothermodule",
+                    "moduleSystem": "cjs",
+                    "valid": true
+                }
+            ]
+        },
+        {
+            "source": "test/cli/fixtures/cjs/node_modules/somemodule/src/moar-javascript.js",
+            "dependencies": []
+        },
+        {
+            "source": "test/cli/fixtures/cjs/node_modules/somemodule/node_modules/someothermodule/main.js",
+            "dependencies": []
+        }
+    ],
+    "summary": {
+        "violations": [],
+        "error": 0,
+        "warn": 0,
+        "info": 0,
+        
+        "optionsUsed": {
+            "outputTo": "test/cli/output/cjs.file.json",
+            "exclude": "",
+            "moduleSystems": [
+                "amd",
+                "cjs",
+                "es6"
+            ],
+            "outputType": "json"
+        }
+    }
+}
