+++ test/report/html/html.spec.js
@@ -1,6 +1,6 @@
 const expect = require("chai").expect;
-const render = require("../../src/report/html");
-const deps = require("./fixtures/cjs-no-dependency-valid.json");
+const render = require("../../../src/report/html");
+const deps = require("./mocks/cjs-no-dependency-valid.json");
 
 const elFixture = `<!DOCTYPE html>
 <html>
